% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_additional_metadata.R
\name{addClass}
\alias{addClass}
\title{add arbitrary S3 class to an object}
\usage{
addClass(x, newClass)
}
\arguments{
\item{x}{an object}

\item{newClass}{character string; class to be added}
}
\value{
The same object with an added S3 class.
}
\description{
Modify and object's \code{class} attribute.
}
\details{
This is a simple convenience function that an item to the \code{class} attribute of an object
so that it can be dispatched to a proper S3 method. This is purely for code clarity,
so that individual methods do not clutter the definitions of higher order functions.
}
\examples{
addClass(data.table::data.table(), "someClass")

}
