% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_to_df.R
\name{assay_to_df}
\alias{assay_to_df}
\title{Transform a SummarizedExperiment assay to a long data.frame.}
\usage{
assay_to_df(se, assay_name, merge_metrics = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment object with dose-response data.}

\item{assay_name}{String of name of the assay in the /code{se}.}

\item{merge_metrics}{Logical indicating whether the metrics should be merged.
Defaults to \code{FALSE}.}
}
\value{
data.frame with dose-response data
}
\description{
Transform a SummarizedExperiment assay to a long data.frame with a single entry for each row and column combination.
}
