% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_choices}
\alias{assert_choices}
\title{assert choices}
\usage{
assert_choices(x, choices, ...)
}
\arguments{
\item{x}{charvec expected subset}

\item{choices}{charvec reference set}

\item{...}{Additional arguments to pass to \code{checkmate::test_choice}}
}
\value{
\code{NULL}
}
\description{
assert choices
}
\examples{
assert_choices("x", c("x","y"))

}
\keyword{package_utils}
