% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_sd}
\alias{calc_sd}
\title{Calculate Standard Deviation or Return Zero}
\usage{
calc_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The standard deviation of the vector if its length is greater than 1 or it is not numeric, otherwise 0.
}
\description{
This function calculates the standard deviation of a numeric vector. If the vector has a length of 1 and it is numeric, it returns 0.
}
\examples{
calc_sd(c(1, 2, 3, 4, 5)) # Should return the standard deviation
calc_sd(c(1)) # Should return 0
calc_sd(numeric(0)) # Should return NA
calc_sd(c("a", "b", "c")) # Should return NA
}
\keyword{package_utils}
