% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{cap_ic50}
\alias{cap_ic50}
\title{Cap IC50 value.}
\usage{
cap_ic50(ic50, max_conc, capping_fold = 5)
}
\arguments{
\item{ic50}{Numeric value of the IC50 to cap.}

\item{max_conc}{Numeric value of the highest concentration in a dose series used to calculate the \code{ic50}.}

\item{capping_fold}{Integer value of the fold number to use for capping.}
}
\value{
Capped IC50 value.
}
\description{
Cap IC50 value by upper and lower limits.
}
\details{
Ideally, the \code{lower_cap} should be based on the lowest tested concentration.
However, since we don't record that, it is set 5 orders of magnitude below the highest dose.
}
