% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_to_dt.R
\name{convert_assay_data_to_dt.default}
\alias{convert_assay_data_to_dt.default}
\title{Convert assay data into data.table (default)}
\usage{
\method{convert_assay_data_to_dt}{default}(object)
}
\arguments{
\item{object}{BumpyDataFrameMatrix or matrix object}
}
\value{
data.table with assay data
}
\description{
Convert assay data into data.table (default)
}
