% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_to_dt.R
\name{convert_assay_data_to_dt.matrix}
\alias{convert_assay_data_to_dt.matrix}
\title{Convert matrix assay data into data.table}
\usage{
\method{convert_assay_data_to_dt}{matrix}(object)
}
\arguments{
\item{object}{matrix object}
}
\value{
data.table with assay data
}
\description{
Convert matrix assay data into data.table
}
