% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{convert_combo_data_to_dt}
\alias{convert_combo_data_to_dt}
\title{convert combo assays from SummarizedExperiments to the list of data.tables}
\usage{
convert_combo_data_to_dt(
  se,
  c_assays = get_combo_assay_names(),
  normalization_type = c("RelativeViability", "GRvalue"),
  prettify = TRUE
)
}
\arguments{
\item{se}{sumamrized experiment with dose-response data}

\item{c_assays}{charvec combo assays to be used}

\item{normalization_type}{charvec normalization_types expected in the data}

\item{prettify}{logic flag prettify the colnames od the returned list of data.tables?}
}
\value{
list of data.table(s) with combo data
}
\description{
convert combo assays from SummarizedExperiments to the list of data.tables
}
\examples{
combo_data_l <- convert_combo_data_to_dt(se)

}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
