% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{convert_combo_field_to_assay}
\alias{convert_combo_field_to_assay}
\title{get combo assay names based on the field name}
\usage{
convert_combo_field_to_assay(field)
}
\value{
charvec
}
\description{
get combo assay names based on the field name
}
\examples{
get_combo_excess_assay_names()

}
