% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_assay_to_dt.R
\name{convert_mae_assay_to_dt}
\alias{convert_mae_assay_to_dt}
\title{Convert a MultiAssayExperiment assay to a long data.table}
\usage{
convert_mae_assay_to_dt(
  mae,
  assay_name,
  experiment_name = NULL,
  include_metadata = TRUE,
  retain_nested_rownames = FALSE
)
}
\arguments{
\item{mae}{A \linkS4class{MultiAssayExperiment} object holding experiments with
raw and/or processed dose-response data in its assays.}

\item{assay_name}{String of name of the assay to transform within the \code{se}.}

\item{experiment_name}{String of name of the experiment in \code{mae} whose \code{assay_name} should be converted.
Default to \code{NULL} that all the experiment should be converted into one data.table object.}

\item{include_metadata}{Boolean indicating whether or not to include \code{rowData(se)}
and \code{colData(se)} in the returned data.table.
Defaults to \code{TRUE}.}

\item{retain_nested_rownames}{Boolean indicating whether or not to retain the rownames
nested within a \code{BumpyMatrix} assay.
Defaults to \code{FALSE}.
If the \code{assay_name} is not of the \code{BumpyMatrix} class, this argument's value is ignored.
If \code{TRUE}, the resulting column in the data.table will be named as \code{"<assay_name>_rownames"}.}
}
\value{
data.table representation of the data in \code{assay_name}.
}
\description{
Convert an assay within a \linkS4class{SummarizedExperiment} object in a MultiAssayExperiment
to a long data.table.
}
\details{
NOTE: to extract information about 'Control' data, simply call the
function with the name of the assay holding data on controls.
To extract the reference data in to same format as 'Averaged' use \code{convert_mae_ref_assay_to_dt}.
}
\examples{
mae <- get_synthetic_data("finalMAE_small.RDS")
convert_mae_assay_to_dt(mae, "Metrics")

}
\seealso{
flatten
}
\author{
Bartosz Czech \href{mailto:bartosz.czech@contractors.roche.com}{bartosz.czech@contractors.roche.com}
}
