% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_assay_to_dt.R
\name{convert_mae_ref_assay_to_dt}
\alias{convert_mae_ref_assay_to_dt}
\title{Convert the reference values from a MultiAssayExperiment assay to a long data.table}
\usage{
convert_mae_ref_assay_to_dt(
  mae,
  experiment_name = NULL,
  ref_relative_viability_assay = "RefRelativeViability",
  ref_gr_value_assay = "RefGRvalue"
)
}
\arguments{
\item{experiment_name}{String of name of the experiment in `mae` whose `assay_name` should be converted.
Default to `NULL` that all the experiment should be converted into one data.table object.}

\item{ref_relative_viability_assay}{String of the name of the assay in the \code{se} 
holding the reference relative viability data.}

\item{ref_gr_value_assay}{String of the name of the assay in the \code{se} holding the reference GR value data.}

\item{se}{A \linkS4class{MultiAssayExperiment} object holding reference data in its assays.}
}
\value{
data.table representation of the reference data.
}
\description{
Transform the Ref[RelativeViability/GRvalue] within a \linkS4class{MultiAssayExperiment} object to a long data.table.
Clean up the column names and add columns to match the format of the data.table from the \code{'Averaged'} assay.
}
\details{
This is a convenience function to massage the reference data into the same format as the \code{"Averaged"}
assay data.
}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
