% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_convert.R
\name{convert_mae_to_json}
\alias{convert_mae_to_json}
\title{Create JSON document.}
\usage{
convert_mae_to_json(mae)
}
\arguments{
\item{mae}{SummarizedExperiment object.}
}
\value{
String representation of a JSON document.
}
\description{
Convert a MultiAssayExperiment object to a JSON document.
}
