% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataframe_to_datatable}
\alias{dataframe_to_datatable}
\title{Convert object to data.table using \code{data.table::setDT}}
\usage{
dataframe_to_datatable(obj)
}
\arguments{
\item{obj}{any object}
}
\value{
a data.table object
}
\description{
Convert object to data.table using \code{data.table::setDT}
}
\examples{
dataframe_to_datatable(list(a = 2)) 

}
\author{
Sergiu Mocanu \href{mailto:sergiu.mocanu@contractors.roche.com}{sergiu.mocanu@contractors.roche.com}
}
