% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatentate_SEs.R
\name{demote_fields}
\alias{demote_fields}
\title{Demote a metadata field to be represented as a nested field of the \code{BumpyMatrix} assay.}
\usage{
demote_fields(se, fields)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object.}

\item{fields}{Character vector of nested fields to demote as nested columns.}
}
\value{
A \code{SummarizedExperiment} object with new dimensions resulting from demoting given \code{fields}.
}
\description{
Demote a metadata field to be represented as a nested field of the \code{BumpyMatrix} assay.
}
\seealso{
promote_fields
}
