% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_convert.R
\name{.convert_colData_to_json}
\alias{.convert_colData_to_json}
\title{Convert colData to JSON}
\usage{
.convert_colData_to_json(cdata, identifiers)
}
\arguments{
\item{cdata}{data.frame of \code{colData}.}
}
\value{
JSON string capturing the \code{cdata}.
}
\description{
Convert colData to JSON format for elasticsearch indexing.
}
\details{
Standardizes the \code{cdata} to common schema fields
and tidies formatting to be condusive to joining 
with other JSON responses.
}
\examples{
cdata <- data.frame(mycellline = letters, mycelllinename = letters, mycelllinetissue = letters)
identifiers <- list(cellline = "mycellline",
                    cellline_name = "mycelllinename",
                    cellline_tissue = "mycelllinetissue")
gDRutils:::.convert_colData_to_json(cdata, identifiers) 

}
\keyword{internal}
