% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_convert.R
\name{.convert_mae_summary_to_json}
\alias{.convert_mae_summary_to_json}
\title{Create JSON document with MAE summary}
\usage{
.convert_mae_summary_to_json(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object.}
}
\value{
String representation of a JSON document.
}
\description{
Create JSON document with MAE summary,
 currently only experiment names
}
