% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{.convert_norm_specific_metrics}
\alias{.convert_norm_specific_metrics}
\title{This function change raw names of metric from long format table into more descriptive
names in the wide format table.
It works for metrics: \code{colnames(get_header("metrics_names"))}}
\usage{
.convert_norm_specific_metrics(x, normalization_type)
}
\description{
This function change raw names of metric from long format table into more descriptive
names in the wide format table.
It works for metrics: \code{colnames(get_header("metrics_names"))}
}
\keyword{internal}
