% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_bm_assay.R
\name{.get_treated_conditions}
\alias{.get_treated_conditions}
\title{.get_treated_conditions}
\usage{
.get_treated_conditions(drug_data)
}
\arguments{
\item{drug_data}{data.table with treatment information}
}
\value{
character vector with treated conditions
}
\description{
Get treated conditions
}
