% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{.metric_output_template}
\alias{.metric_output_template}
\title{Build a named list of NA values for all response metric columns.}
\usage{
.metric_output_template()
}
\description{
Build a named list of NA values for all response metric columns.
}
\keyword{internal}
