% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prep_cd_conc_cap_dict}
\alias{.prep_cd_conc_cap_dict}
\title{Prepare dict with min and max concentration for codilution}
\usage{
.prep_cd_conc_cap_dict(conc_assay_dt, group_cols)
}
\arguments{
\item{conc_assay_dt}{assay data in data.table format with Concentration data}

\item{group_cols}{charvec with grouping column names}
}
\value{
\code{data.table} with max and min concentration for codilution
}
\description{
Prepare dict with min and max concentration for codilution
}
\keyword{internal}
