% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{.snap_conc_to_model}
\alias{.snap_conc_to_model}
\title{Snap a concentration to the nearest available model concentration}
\usage{
.snap_conc_to_model(user_conc, available_concs)
}
\arguments{
\item{user_conc}{A single numeric value for the desired concentration.}

\item{available_concs}{A numeric vector of concentrations for which a model exists.}
}
\value{
A single numeric value from 'available_concs'.
}
\description{
Finds the value in a vector of available concentrations that is closest
(on a log scale) to a user-specified concentration. This is an internal
helper function.
}
\keyword{internal}
