% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{dump_config}
\alias{dump_config}
\title{A function that creates a configuration dump}
\usage{
dump_config(config, tab = "  ", eol = "\\n", level = 0)
}
\arguments{
\item{config}{config (R list)}

\item{tab}{tab (prefix) string}

\item{eol}{end of line string}

\item{level}{current config level (for internal use)}
}
\value{
string "tree" representation of the config
}
\description{
A function that creates a configuration dump
}
