% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{evaluate_efficacy_from_conc}
\alias{evaluate_efficacy_from_conc}
\title{Evaluate efficacy values given fit parameters and a concentration.}
\usage{
evaluate_efficacy_from_conc(c, Vinf, V0, EC50, h)
}
\description{
Evaluate efficacy values given fit parameters and a concentration.
}
\details{
The inverse of this function is \code{predict_conc_from_fit}.
}
\seealso{
predict_conc_from_fit
}
