% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_assay_to_dt.R
\name{flatten_stacked_dt}
\alias{flatten_stacked_dt}
\title{Flatten a data.table}
\usage{
flatten_stacked_dt(dt, columns, flatten, sep = "_")
}
\arguments{
\item{dt}{data.table to flatten.}

\item{columns}{character vector of column names representing uniquifying groups in expansion.}

\item{flatten}{character vector of column names to flatten.}

\item{sep}{string representing separator between \code{flatten} columns, used in column renaming.
Defaults to \code{"_"}.}
}
\value{
data.table of flattened data as defined by \code{flatten}.
}
\description{
Flatten a stacked data.table into a wide format.
}
\details{
flattened columns will be named with original column names prefixed by \code{flatten} columns,
concatenated together and separated by \code{sep}.

A common use case for this function is when a flattened version of the \code{"Metrics"} assay is desired.
}
\examples{
 n <- 4
 m <- 5
 grid <- expand.grid(normalization_type = c("GR", "RV"),
   source = c("GDS", "GDR"))
 repgrid <- do.call("rbind", rep(list(grid), m))
 repgrid$wide <- seq(m * n)
 repgrid$id <- rep(LETTERS[1:m], each = n)

 columns <- colnames(grid)
 flatten <- c("wide")

 flatten_stacked_dt(repgrid, columns = columns, flatten = flatten)

}
