% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_assay_to_dt.R
\name{friendly_names}
\alias{friendly_names}
\title{Return friendly names for metrics}
\usage{
friendly_names(name_list, gDRviz_format = F)
}
\arguments{
\item{name_list}{a list of names.}

\item{format}{either \code{"variable"} (default) or \code{"gDRviz"}.}
}
\value{
table of data with updated column names.
}
\description{
Return friendly names for metrics.
}
\details{
Rename names that are  metrics in a table.

A common use case for this function is to convert column names from a flattened version of the \code{"Metrics"} assay.
}
