% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\docType{package}
\name{gDRutils}
\alias{gDRutils}
\alias{.datatable.aware}
\title{'gDRutils - a package with utils'}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
A package with helper functions for processing drug response data.
}
\keyword{datasets}
