% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tests.R
\name{gen_synthetic_data}
\alias{gen_synthetic_data}
\title{gen_synthetic_data}
\usage{
gen_synthetic_data(m = 1, n = 1)
}
\value{
list with drugs, cell_lines, raw_data and assay_data
}
\description{
Function for generating local synthetic data used for unit tests in modules
}
\examples{
gen_synthetic_data()

}
