% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tests.R
\name{gen_testdata}
\alias{gen_testdata}
\title{gen_testdata}
\usage{
gen_testdata()
}
\value{
list with drugs, cell_lines, raw_data and assay_data
}
\description{
Function to obtain data from gDRtestData and prepare for unit tests
}
\examples{
gen_testdata()

}
