% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean}
\usage{
geometric_mean(x, fixed = TRUE, maxlog10Concentration = 1)
}
\arguments{
\item{x}{numeric vector}

\item{fixed}{flag should be add fix for -Inf}

\item{maxlog10Concentration}{numeric value needed to calculate minimal value}
}
\value{
numeric vector
}
\description{
Auxiliary function for calculating geometric mean with possibility to handle -Inf
}
\keyword{internal}
