% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_names.R
\name{get_assay_names}
\alias{get_assay_names}
\title{get assay names of the given se/dataset
fetch the data from the se if provided as metadata
use predefined values from \code{get_env_assay_names} otherwise}
\usage{
get_assay_names(se = NULL, ...)
}
\arguments{
\item{se}{SummarizedExperiment or NULL}

\item{...}{Additional arguments to pass to \code{get_env_assay_names}.}
}
\value{
charvec
}
\description{
get assay names of the given se/dataset
fetch the data from the se if provided as metadata
use predefined values from \code{get_env_assay_names} otherwise
}
\examples{
get_assay_names()

}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
