% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{get_assay_uniq_cols}
\alias{get_assay_uniq_cols}
\title{get columns in the assay data required to have unique data}
\usage{
get_assay_uniq_cols(dt)
}
\arguments{
\item{dt}{data.table with assay data}
}
\value{
charvec with columns required to have unique data
}
\description{
get columns in the assay data required to have unique (non-duplicated) data
}
\examples{
sdata <- get_synthetic_data("finalMAE_small")
smetrics_data <- convert_se_assay_to_dt(sdata[[1]], "Metrics")
get_assay_uniq_cols(smetrics_data)
}
\keyword{duplicates}
