% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{get_combo_assay_names}
\alias{get_combo_assay_names}
\title{get names of combo assays}
\usage{
get_combo_assay_names(se = NULL, ...)
}
\value{
charvec
}
\description{
get names of combo assays
}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
