% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_names.R
\name{get_combo_base_assay_names}
\alias{get_combo_base_assay_names}
\title{get names of combo base assays}
\usage{
get_combo_base_assay_names(se = NULL, ...)
}
\arguments{
\item{se}{SummarizedExperiment or NULL}

\item{...}{Additional arguments to pass to \code{get_combo_assay_names}.}
}
\value{
charvec
}
\description{
get names of combo base assays
}
\examples{
get_combo_base_assay_names()
}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
