% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{get_combo_col_settings}
\alias{get_combo_col_settings}
\title{Get colorscale data for given combo assay and growth metric}
\usage{
get_combo_col_settings(g_metric, assay_type)
}
\arguments{
\item{g_metric}{growth metric}

\item{assay_type}{assay_type}
}
\value{
list with colors, breaks and limits
}
\description{
Get colorscale data for given combo assay and growth metric
}
\examples{

}
