% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{get_combo_excess_field_names}
\alias{get_combo_excess_field_names}
\title{get names of combo excess fields}
\usage{
get_combo_excess_field_names()
}
\value{
charvec
}
\description{
get names of combo excess fields
}
\examples{
get_combo_excess_field_names()

}
