% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{get_combo_score_field_names}
\alias{get_combo_score_field_names}
\title{get names of combo score fields}
\usage{
get_combo_score_field_names()
}
\value{
charvec
}
\description{
get names of combo score fields
}
\examples{
get_combo_score_assay_names()

}
\keyword{combination_data}
