% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{get_config_item}
\alias{get_config_item}
\title{Get config element by name}
\usage{
get_config_item(config, path, default = NULL, stop_on_missing = FALSE)
}
\arguments{
\item{config}{config object (R list)}

\item{path}{string or character vector containing the path to the config item}

\item{default}{default value for missing item}

\item{stop_on_missing}{call "stop" if required item is missing}
}
\value{
config item
}
\description{
Get config element by name
}
