% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_default_identifiers}
\alias{get_default_identifiers}
\title{Get gDR default identifiers required for downstream analysis.}
\usage{
get_default_identifiers()
}
\description{
Get gDR default identifiers required for downstream analysis.
}
