% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_names.R
\name{get_env_assay_names}
\alias{get_env_assay_names}
\title{get default assay names
for the specified filters,
i.e. set of assay types, assay groups and assay data types}
\usage{
get_env_assay_names(
  type = NULL,
  group = NULL,
  data_type = NULL,
  prettify = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{type}{charvec of assay types}

\item{group}{charvec of assay groups}

\item{data_type}{charvec assay of data types}

\item{prettify}{logical flag, prettify the assay name?}

\item{simplify}{logical flag, simplify the output?
will return single string instead of named vector with single element
useful when function is expected to return single element/assay only}
}
\description{
get default assay names
for the specified filters,
i.e. set of assay types, assay groups and assay data types
}
\author{
Arkadiusz Gładki \email{arkadiusz.gladki@contractors.roche.com}
}
