% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{get_env_configs}
\alias{get_env_configs}
\title{Get config files from environment variable(s)}
\usage{
get_env_configs(prefix = "GDR_CONFIG")
}
\arguments{
\item{prefix}{beginning of the name of the environment variables}
}
\value{
NULL or character vector of config files
}
\description{
Get config files from environment variable(s)
}
