% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_env_var}
\alias{get_env_var}
\title{safe wrapper of Sys.getenv()}
\usage{
get_env_var(x, ...)
}
\arguments{
\item{x}{string with the name of the environmental variable}

\item{...}{additional params for Sys.getenev}
}
\value{
sanitized value of the env variable
}
\description{
So far the helper is needed to handle env vars containing \code{:}
for which the backslash  is automatically added in some contexts
and R could not get the original value for these env vars.
}
\examples{
get_env_var("HOME")
}
\keyword{package_utils}
