% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_expect_one_identifiers}
\alias{get_expect_one_identifiers}
\title{Get identifiers that expect only one value for each identifier.}
\usage{
get_expect_one_identifiers()
}
\value{
charvec
}
\description{
Get identifiers that expect only one value for each identifier.
}
\examples{
get_expect_one_identifiers()

}
\keyword{identifiers}
