% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_experiment_groups}
\alias{get_experiment_groups}
\title{get_experiment_groups}
\usage{
get_experiment_groups(type = NULL)
}
\arguments{
\item{type}{String indicating the name of an assay group.
Defaults to all experiment groups.}
}
\value{
list with experiment groups or string (if type not NULL)
}
\description{
get experiment groups
}
\examples{
get_experiment_groups()

}
\author{
Arkadiusz Gladki <arkadiusz.gladki@contractors.roche.com>
}
