% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gDR_session_info}
\alias{get_gDR_session_info}
\title{get gDR package and their version installed in the environment}
\usage{
get_gDR_session_info(pattern = "^gDR")
}
\arguments{
\item{pattern}{string with the pattern to grep R packages from the list of installed packages}
}
\value{
data.table with gDR packages and their versions
}
\description{
get gDR package and their version installed in the environment
}
\examples{
get_gDR_session_info()

}
\keyword{package_utils}
