% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_identifier}
\alias{get_identifier}
\title{Deprecated function, use get_env_identifier.}
\usage{
get_identifier(k = NULL, simplify = TRUE)
}
\arguments{
\item{k}{String corresponding to identifier name.}

\item{simplify}{Boolean indicating whether output should be simplified.}
}
\description{
Deprecated function, use get_env_identifier.
}
