% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_identifiers_dt}
\alias{get_identifiers_dt}
\title{Get descriptions for identifiers}
\usage{
get_identifiers_dt(k = NULL, get_description = FALSE, get_example = FALSE)
}
\arguments{
\item{k}{identifier key, string}

\item{get_description}{return descriptions only, boolean}

\item{get_example}{return examples only, boolean}
}
\description{
Get descriptions for identifiers
}
