% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_idfs_synonyms}
\alias{get_idfs_synonyms}
\title{Get gDR synonyms for the identifiers}
\usage{
get_idfs_synonyms()
}
\value{
charvec
}
\description{
Get gDR synonyms for the identifiers
}
\examples{
get_idfs_synonyms()

}
\keyword{identifiers}
