% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{get_iso_colors}
\alias{get_iso_colors}
\title{get_iso_colors
#'}
\usage{
get_iso_colors(normalization_type = c("RelativeViability", "GRvalue"))
}
\arguments{
\item{normalization_type}{charvec normalization_types expected in the data}
}
\value{
named charvec with iso colors
}
\description{
get_iso_colors
#'
}
