% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_const.R
\name{get_isobologram_columns}
\alias{get_isobologram_columns}
\title{Get isobologram column names}
\usage{
get_isobologram_columns(k = NULL, prettify = TRUE)
}
\arguments{
\item{k}{key}

\item{prettify}{change to upper case and add underscore, iso_level --> Iso_Level}
}
\value{
character vector of isobologram column names for combination data
}
\description{
Get isobologram column names
}
\examples{
get_isobologram_columns()
get_isobologram_columns("iso_level", prettify = TRUE)

}
\keyword{json_const}
