% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_validators.R
\name{get_optional_experiment_field}
\alias{get_optional_experiment_field}
\title{Get Optional Experiment Field}
\usage{
get_optional_experiment_field()
}
\value{
a list containing the names of the optional identifiers and the default value
}
\description{
Get Optional Experiment Field
}
\author{
Sergiu Mocanu <sergiu.mocanu@contractors.roche.com>
}
