% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{get_path_configs}
\alias{get_path_configs}
\title{Get config files from directory}
\usage{
get_path_configs(dir)
}
\arguments{
\item{dir}{path to a directory which may contain configuration files}
}
\value{
NULL or character vector with config files
}
\description{
Get config files from directory
}
