% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_required_identifiers}
\alias{get_required_identifiers}
\title{Get identifiers required for downstream analysis.}
\usage{
get_required_identifiers()
}
\value{
charvec
}
\description{
Get identifiers required for downstream analysis.
}
\examples{
get_required_identifiers()

}
\keyword{identifiers}
