% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_metadata.R
\name{get_SE_fit_parameters}
\alias{get_SE_fit_parameters}
\title{Get metadata for fit parameters metadata in SummarizedExperiment object.}
\usage{
get_SE_fit_parameters(se)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to get fit parameters.}
}
\value{
named list of fitting parameters.
}
\description{
Get metadata for fit parameters used to construct the Metrics assay in a SummarizedExperiment object.
}
