% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_metadata.R
\name{get_SE_processing_metadata}
\alias{get_SE_processing_metadata}
\title{Set metadata for the processing info in SummarizedExperiment object.}
\usage{
get_SE_processing_metadata(se)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to get processing info.}
}
\value{
named list of processing info.
}
\description{
Set metadata for the processing info that defines 
the date_processed and packages versions in SummarizedExperiment object metadata.
}
