% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_const.R
\name{get_settings_from_json}
\alias{get_settings_from_json}
\title{Get settings from JSON file
In most common scenario the settings are stored in JSON file
to avoid hardcoding}
\usage{
get_settings_from_json(
  s = NULL,
  json_path = system.file(package = "gDRutils", "cache.json")
)
}
\arguments{
\item{s}{charvec with setting entry/entries}

\item{json_path}{string with the path to the JSON file}
}
\value{
value/values for entry/entries from JSON file
}
\description{
Get settings from JSON file
In most common scenario the settings are stored in JSON file
to avoid hardcoding
}
\examples{
\dontrun{
get_settings_from_json()
}

}
\keyword{json_const}
