% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tests.R
\name{get_testdata}
\alias{get_testdata}
\title{get_testdata}
\usage{
get_testdata()
}
\value{
list with drugs, cell_lines, raw_data and assay_data
}
\description{
Function to obtain data from gDRtestData and prepare for unit tests
}
\examples{
get_testdata()

}
