% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-tests.R
\name{get_testdata_combo}
\alias{get_testdata_combo}
\title{get_testdata_combo}
\usage{
get_testdata_combo()
}
\value{
list with drugs, cell_lines, raw_data and assay_data
}
\description{
Function to obtain data from gDRtestData and prepare for unit tests
}
\examples{
get_testdata_combo()

}
\keyword{test_helpers}
