% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{has_dt_duplicated_rows}
\alias{has_dt_duplicated_rows}
\title{check if data.table contains duplicated data}
\usage{
has_dt_duplicated_rows(dt, col_names = NULL)
}
\arguments{
\item{dt}{data.table}

\item{col_names}{charvec with columns to be used for subsetting}
}
\value{
logical flag
}
\description{
An auxiliary function that checks for duplicates in the data.table (or its subset)
}
\examples{
dt <- data.table::data.table(a = c(1, 2, 3), b = c(3, 2, 2))
has_dt_duplicated_rows(dt, "b")
}
\keyword{duplicates}
