% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_single_codrug_data}
\alias{has_single_codrug_data}
\title{Has Single Codrug Data}
\usage{
has_single_codrug_data(
  cols,
  prettify_identifiers = TRUE,
  codrug_identifiers = c("drug_name2", "concentration2")
)
}
\arguments{
\item{cols}{character vector with the columns of the input data}

\item{prettify_identifiers}{logical flag specifying if identifiers are expected to be prettified}

\item{codrug_identifiers}{character vector with identifiers for the codrug columns}
}
\value{
logical flag
}
\description{
Has Single Codrug Data
}
\examples{
has_single_codrug_data("Drug Name")
has_single_codrug_data(c("Drug Name", "Cell Lines"))
has_single_codrug_data(c("Drug Name 2", "Concentration 2"))
has_single_codrug_data(
  get_prettified_identifiers(
    c("concentration2", "drug_name2"),
    simplify = FALSE
  )
)

}
\keyword{combination_data}
