% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_any_exp_empty}
\alias{is_any_exp_empty}
\title{is_any_exp_empty}
\usage{
is_any_exp_empty(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
logical
}
\description{
check if any experiment is empty
}
\examples{
mae <- get_synthetic_data("finalMAE_small.RDS")
is_any_exp_empty(mae)

}
\author{
Arkadiusz Gladki <arkadiusz.gladki@contractors.roche.com>
}
