% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_combo_data}
\alias{is_combo_data}
\title{Checks if \code{se} is combo dataset.}
\usage{
is_combo_data(se)
}
\arguments{
\item{se}{SummarizedExperiment}
}
\value{
logical
}
\description{
Checks if \code{se} is combo dataset.
}
\examples{
se <- get_synthetic_data("finalMAE_combo_matrix.qs2")[[1]]
is_combo_data(se)
se <- get_synthetic_data("finalMAE_combo_matrix.qs2")[[2]]
is_combo_data(se)
se <- get_synthetic_data("finalMAE_small.qs2")[[1]]
is_combo_data(se)

}
\keyword{combination_data}
