% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_exp_empty}
\alias{is_exp_empty}
\title{is_exp_empty}
\usage{
is_exp_empty(exp)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
logical
}
\description{
check if experiment (SE) is empty
}
\author{
Arkadiusz Gladki <arkadiusz.gladki@contractors.roche.com>
}
