% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{load_configuration}
\alias{load_configuration}
\title{Load configuration}
\usage{
load_configuration(
  ...,
  use_default = TRUE,
  replacements = get_replacements(),
  evaluate = list(R = FALSE, PKG = TRUE, ENV = TRUE),
  dump = FALSE
)
}
\arguments{
\item{...}{optional paths to configuration files and/or paths to directories that may contain these files}

\item{use_default}{load default.yaml config from this package}

\item{replacements}{named list containing replacements e.g list(from = 'to something new')}

\item{evaluate}{list with flags responsible for evaluations R, PKG, ENV strings}

\item{dump}{flag responsible for displaying the contents of the final configuration}
}
\value{
Config "object" (R list)
}
\description{
Load configuration
}
\examples{
\dontrun{
--- example config files ---

gDRutils/config/default.yaml
  default: test

a.yaml
  x1: a
  x2:
    x21: a
    x22: a
  x3: ["a1" , "a2"]
  r1: "{from_str_1}"
  r2: "{from_str_2}"

b.yaml
  x1: b
  x2:
    x21: b
    x23: b
  e1: "`PKG:gDRviz`/subdir"
  e2: "`ENV:X`"
  e2: `R:paste0("evaluated")`

-- R code --

load_configuration(
 "a.yaml",
 repacements = list(from_str_1 = "to str 1", from_str_3 = "doesn't exist")
)
  default: test
  x1: a
  x2:
    x21: a
    x22: a
  x3: ["a1" , "a2"]
  r1: "to_str_1"
  r2: "{from_str_2}"

load_configuration(
  "a.yaml", "b.yaml", use_default = FALSE,
  evaluate = list(PKG = TRUE, ENV = FALSE, R = TRUE)
)
  x1: b
  x2:
    x21: b
    x22: a
    x23: b
  x3: ["a1" , "a2"]
  r1: "{from_str_1}"
  r2: "{from_str_2}"
  e1: "/usr/local/lib/R/library/gDRviz/subdir"
  e2: "`ENV:X`"
  e2: evaluated

Sys.setenv(X = "X value")
set_replacements(from_str_2 = "to two")
load_configuration(
  "a.yaml", "b.yaml",
  evaluate = list(PKG = FALSE, ENV = TRUE, R = FALSE)
)
  default: test
  x1: b
  x2:
    x21: b
    x22: a
    x23: b
  x3: ["a1" , "a2"]
  r1: "{from_str_1}"
  r2: "to_two"
  e1: "`PKG:gDRviz`/subdir"
  e2: "X value"
  e2: `R:paste0("evaluated")`
}

}
