% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{logistic_4parameters}
\alias{logistic_4parameters}
\title{Deprecated function, use predict_efficacy_from_conc.}
\usage{
logistic_4parameters(c, Vinf, V0, EC50, h)
}
\arguments{
\item{c}{Numeric vector representing concentrations to predict efficacies for.}

\item{Vinf}{Numeric vector representing the asymptotic value of the sigmoidal fit to the dose-response
data as concentration goes to infinity.}

\item{V0}{Numeric vector representing the asymptotic metric value corresponding to a concentration of 0
for the primary drug.}

\item{EC50}{Numeric vector representing the drug concentration at half-maximal effect.}

\item{h}{Numeric vector representing the hill coefficient of the fitted curve, which reflects how steep
the dose-response curve is.}
}
\description{
Deprecated function, use predict_efficacy_from_conc.
}
