% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MAEpply}
\alias{MAEpply}
\title{Lapply through all the experiments in MultiAssayExperiment object}
\usage{
MAEpply(mae, FUN, unify = FALSE, ...)
}
\arguments{
\item{mae}{MultiAssayExperiment object}

\item{FUN}{function that should be applied on each experiment of MultiAssayExperiment object}

\item{unify}{logical indicating if the output should be a unlisted object of unique
values across all the experiments}

\item{...}{Additional args to be passed to teh \code{FUN}.}
}
\description{
Lapply through all the experiments in MultiAssayExperiment object
}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
