% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcolData}
\alias{mcolData}
\title{mcolData}
\usage{
mcolData(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
tibble with all-experiments colData
}
\description{
get colData of all experiments
}
\author{
Arkadiusz Gladki <arkadiusz.gladki@contractors.roche.com>
}
