% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{merge_configs}
\alias{merge_configs}
\title{Merge two "config" lists}
\usage{
merge_configs(base_config, overlay_config, recursive = TRUE)
}
\arguments{
\item{base_config}{base config list}

\item{overlay_config}{overlay config list}

\item{recursive}{merge recursive flag}
}
\value{
merged config lists
}
\description{
Merge two "config" lists
}
