% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge two "config" lists}
\usage{
merge_lists(base_list, overlay_list, recursive = TRUE)
}
\arguments{
\item{base_list}{base list}

\item{overlay_list}{overlay list}

\item{recursive}{merge recursive flag}
}
\value{
merged config list
}
\description{
Merge two "config" lists
}
