% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_SE.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge metadata}
\usage{
merge_metadata(SElist, metadata_fields)
}
\arguments{
\item{SElist}{named list of \code{SummarizedExperiment}s}

\item{metadata_fields}{vector of metadata names that will be merged}
}
\value{
list of merged metadata
}
\description{
Merge metadata
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs2")
se <- mae[[1]]
listSE <- list(
  se,
  se
)
metadata_fields <- identify_unique_se_metadata_fields(listSE)
merge_metadata(listSE, metadata_fields)

}
\keyword{SE_operators}
