% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_SE.R
\name{merge_SE}
\alias{merge_SE}
\title{Merge multiple Summarized Experiments}
\usage{
merge_SE(SE)
}
\arguments{
\item{SE}{list of Summarized Experiments}
}
\value{
Summarized Experiment
}
\description{
Merge multiple Summarized Experiments
}
