% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mrowData}
\alias{mrowData}
\title{mrowData}
\usage{
mrowData(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
data.table with all-experiments rowData
}
\description{
get rowData of all experiments
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs2")
mrowData(mae)

}
\author{
Arkadiusz Gladki \href{mailto:arkadiusz.gladki@contractors.roche.com}{arkadiusz.gladki@contractors.roche.com}
}
\keyword{package_utils}
