% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{notify_if_duplicates}
\alias{notify_if_duplicates}
\title{send email if assay data.table contains duplicated rows}
\usage{
notify_if_duplicates(
  dt,
  by = "email",
  assay_name = "unknown",
  preview_max_numb = 4,
  metadata = NULL
)
}
\arguments{
\item{dt}{data.table with assay data}

\item{by}{charvec with notification methods (currently 'email' supported)}

\item{assay_name}{string with the name of the assay}

\item{preview_max_numb}{number of rows to preview if duplicates found}

\item{metadata}{list with the additional metadata to send via email}
}
\description{
An auxiliary function to send email if duplicated rows in assay data.table are found
}
\examples{
sdata <- get_synthetic_data("finalMAE_small")
smetrics_data <- convert_se_assay_to_dt(sdata[[1]], "Metrics")
throw_msg_if_duplicates(smetrics_data, assay_name = "Metrics", msg_f = futile.logger::flog.info)
}
\keyword{duplicates}
