% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{predict_conc_from_efficacy}
\alias{predict_conc_from_efficacy}
\title{Predict a concentration for a given efficacy with fit parameters.}
\usage{
predict_conc_from_efficacy(efficacy, x_inf, x_0, ec50, h)
}
\description{
Predict a concentration for a given efficacy with fit parameters.
}
\details{
The inverse of this function is \code{predict_efficacy_from_conc}.
}
\seealso{
predict_efficacy_from_conc .calculate_x50
}
