% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{predict_conc_from_fit}
\alias{predict_conc_from_fit}
\title{Predict a concentration for a given efficacy from a fit.}
\usage{
predict_conc_from_fit(ec50, x0, xInf, h, efficacy)
}
\description{
Predict a concentration for a given efficacy from a fit.
}
\details{
The inverse of this function is \code{logistic_4parameters}.
}
\seealso{
evaluate_efficacy_from_conc .calculate_x50
}
