% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{predict_efficacy_from_conc}
\alias{predict_efficacy_from_conc}
\title{Predict efficacy values given fit parameters and a concentration.}
\usage{
predict_efficacy_from_conc(c, x_inf, x_0, ec50, h)
}
\description{
Predict efficacy values given fit parameters and a concentration.
}
\details{
The inverse of this function is \code{predict_conc_from_efficacy}.
}
\seealso{
predict_conc_from_efficacy
}
