% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatentate_SEs.R
\name{promote_fields}
\alias{promote_fields}
\title{Promote a nested field to be represented as a metadata field of the \code{SummarizedExperiment}
as either the \code{rowData} or \code{colData}.}
\usage{
promote_fields(se, fields, MARGIN = c(1, 2))
}
\arguments{
\item{se}{\code{SummarizedExperiment} object.}

\item{fields}{Character vector of nested fields to promote.}

\item{MARGIN}{Numeric of values \code{1} or \code{2} indicating whether to 
promote fields to rows or columns respectively.}
}
\value{
A \code{SummarizedExperiment} object with new dimensions resulting from promoting given \code{fields}.
}
\description{
Promote a nested field to be represented as a metadata field of the \code{SummarizedExperiment}
as either the \code{rowData} or \code{colData}.
}
\seealso{
demote_fields
}
