% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{read_config}
\alias{read_config}
\title{Function that read the configuration from the yaml file}
\usage{
read_config(path, as_yaml = FALSE)
}
\arguments{
\item{path}{path to the yaml file}

\item{as_yaml}{encode yaml content flag}
}
\value{
NULL if error occurred, string if as_yaml = FALSE, otherwise config object
}
\description{
Function that read the configuration from the yaml file
}
