% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{refine_coldata}
\alias{refine_coldata}
\title{refine colData}
\usage{
refine_coldata(cd, se, default_v = "Undefined")
}
\arguments{
\item{cd}{DataFrame with colData}

\item{se}{a SummarizedExperiment object with drug-response data generate by gDR pipeline}

\item{default_v}{string with default value for optional columns in colData}
}
\value{
refined colData
}
\description{
current improvements done on the colData as a standardization step:
\itemize{
\item set default value for optional colData fields
}
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
refine_coldata(SummarizedExperiment::colData(mae[[1]]), mae[[1]])

}
