% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{rename_DFrame}
\alias{rename_DFrame}
\title{Rename DFrame}
\usage{
rename_DFrame(df, mapping_vector)
}
\arguments{
\item{df}{a DFrame object}

\item{mapping_vector}{a named vector for mapping old and new values.
The names of the character vector indicate the source names, and the
corresponding values the destination names.}
}
\value{
a renamed DFrame object
}
\description{
Rename DFrame
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
rename_DFrame(SummarizedExperiment::rowData(mae[[1]]), c("Gnumber" = "Gnumber1"))

}
