% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{RVGRfits}
\alias{RVGRfits}
\title{Actual fitting function}
\usage{
RVGRfits(
  df_,
  e_0 = 1,
  GR_0 = 1,
  n_point_cutoff = 4,
  range_conc = c(0.005, 5),
  force = FALSE
)
}
\arguments{
\item{e_0}{Defaults to \code{1}.}

\item{GR_0}{Defaults to \code{1}.}

\item{force}{use signifcance or not}

\item{log10concs}{concentrations}

\item{RelativeViability}{values}

\item{GRvalues}{values}

\item{cap}{enforce e_0 and GR_0}
}
\value{
vector of parameters
}
\description{
\code{RVGRfits} returns fit parameters
}
\details{
returns fit parameters
}
