% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{send_email}
\alias{send_email}
\title{send email by wrapping \code{mailR::send.mail}}
\usage{
send_email(
  html_data,
  to,
  subject,
  from = get_env_var("EMAIL_SENDER", "gdr-support-d@gene.com"),
  inline = FALSE
)
}
\arguments{
\item{to}{string with the recipient}

\item{subject}{string with the subject}

\item{from}{string with the sender}

\item{inline}{A boolean indicating whether images in the HTML file should be embedded inline}

\item{body}{body content of the email}

\item{html}{A boolean indicating whether the body of the email should be parsed as HTML}

\item{attached_files}{A character with file paths to be attached to the email}
}
\description{
send email by wrapping \code{mailR::send.mail}
}
\keyword{utils}
