% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identifier.R
\name{set_identifier}
\alias{set_identifier}
\title{Set the identifier(s) for an identifier field}
\usage{
set_identifier(x, v)
}
\arguments{
\item{x}{string of field to set identifiers for}

\item{v}{character vector of identifiers corresponding to identifier field \code{x}}
}
\value{
a \code{NULL} invisibly
}
\description{
Set the expected identifier(s) for a specific identifier field
}
\details{
If \code{set_identifier} is called on an identifier that is already set, an error is thrown.
}
