% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{set_replacements}
\alias{set_replacements}
\title{Set replacements}
\usage{
set_replacements(...)
}
\arguments{
\item{...}{replacements as name = value parameters}
}
\description{
Set replacements
}
