% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_metadata.R
\name{set_SE_experiment_metadata}
\alias{set_SE_experiment_metadata}
\title{Set experiment-level metadata for a SummarizedExperiment object.}
\usage{
set_SE_experiment_metadata(se, value)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to add experiment-level metadata.}

\item{value}{named list of metadata for the \code{se} object.}
}
\value{
\linkS4class{se} with added metadata.
}
\description{
Set experiment-level metadata in the metadata slot of a \linkS4class{SummarizedExperiment} object.
}
