% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_metadata.R
\name{set_SE_fit_parameters}
\alias{set_SE_fit_parameters}
\title{Set metadata for parameters used in fitting in SummarizedExperiment object.}
\usage{
set_SE_fit_parameters(se, value)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to add fit parameter metadata.}

\item{value}{named list of metadata for fit parameters.}
}
\value{
\linkS4class{se} with added metadata.
}
\description{
Set metadata for the fitting parameters that define the Metrics assay in SummarizedExperiment object metadata.
}
