% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_metadata.R
\name{set_SE_processing_info}
\alias{set_SE_processing_info}
\title{Set metadata for the processing info in SummarizedExperiment object.}
\usage{
set_SE_processing_info(se, value)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which 
to add processing info metadata.}

\item{value}{named list of metadata for processing info.}
}
\value{
\code{se} with added metadata.
}
\description{
Set metadata for the processing info that defines 
the date_processed and packages versions in SummarizedExperiment object metadata.
}
