% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shorten_normalization_type_name}
\alias{shorten_normalization_type_name}
\title{shorten normalization type}
\usage{
shorten_normalization_type_name(x)
}
\arguments{
\item{x}{string with normalization type}
}
\value{
shortened string representing the normalization type
}
\description{
shorten normalization type
}
\examples{
shorten_normalization_type_name("GRvalue")

}
\keyword{package_utils}
