% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{standardize_se}
\alias{standardize_se}
\title{Standardize SE by switching from custom identifiers into gDR-default}
\usage{
standardize_se(se)
}
\arguments{
\item{se}{a SummarizedExperiment object with drug-response data generate by gDR pipeline}
}
\value{
se a SummarizedExperiment with default gDR identifiers
}
\description{
Standardize SE by switching from custom identifiers into gDR-default
}
\examples{
se <- get_synthetic_data("finalSE_small.RDS")
standardize_se(se)

}
