% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{str_eval}
\alias{str_eval}
\title{Evaluate `PKG/ENV/R : ...` expressions in string}
\usage{
str_eval(s, evaluate, s_name = paste0("'", s, "'"))
}
\arguments{
\item{s}{string that may contain `PKG/ENV/R : ...` expressions}

\item{evaluate}{list with flags responsible for evaluations R, PKG, ENV strings}

\item{s_name}{friendly name of s string that will be used in log messages}
}
\value{
evaluated string
}
\description{
Evaluate `PKG/ENV/R : ...` expressions in string
}
