% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{str_replace}
\alias{str_replace}
\title{Replace {x} with values defined in replacements}
\usage{
str_replace(s, replacements)
}
\arguments{
\item{s}{string that may contain replacements}

\item{replacements}{named list containing replacements e.g list(from = 'to something new')}
}
\value{
replaced string
}
\description{
Replace {x} with values defined in replacements
}
