% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{update_env_idfs_from_mae}
\alias{update_env_idfs_from_mae}
\title{Update environment identifiers from MAE object identifiers}
\usage{
update_env_idfs_from_mae(mae_idfs)
}
\arguments{
\item{mae_idfs}{A list containing MAE identifiers}
}
\value{
\code{NULL}
}
\description{
Update environment identifiers from MAE object identifiers
}
\examples{
update_env_idfs_from_mae(list(get_env_identifiers()))

}
\keyword{identifiers}
