% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{user_config_file}
\alias{user_config_file}
\title{Helper that creates path to user config file by user config dir and username}
\usage{
user_config_file(username, users_settings_dir)
}
\arguments{
\item{username}{user name}

\item{users_settings_dir}{path to directory where user setting file is (or will be) stored}
}
\value{
path to user config file
}
\description{
Helper that creates path to user config file by user config dir and username
}
