% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_validators.R
\name{validate_dimnames}
\alias{validate_dimnames}
\title{Validate dimnames}
\usage{
validate_dimnames(obj, obj2, skip_empty = TRUE)
}
\arguments{
\item{obj}{first object with dimnames to compare}

\item{obj2}{second object with dimnames to compare}

\item{skip_empty}{a logical indicating whether to skip comparison
if a given dimname is empty in the case of both objects}
}
\value{
\code{NULL}
}
\description{
Assure that dimnames of two objects are the same
}
\keyword{experiment}
\keyword{internal}
