% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_validate.R
\name{validate_json}
\alias{validate_json}
\title{Validate JSON against a schema.}
\usage{
validate_json(json, schema_path)
}
\arguments{
\item{json}{String of JSON in memory.}

\item{schema_path}{String of the schema to validate against.}
}
\value{
Boolean of whether or not JSON successfully validated.
}
\description{
Validate JSON describing an object against a schema.
}
\details{
This is most often used to validate JSON
before passing it in as a document to an ElasticSearch index.
}
\examples{
json <- '{}'

}
\keyword{json_validate}
