% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_validators.R
\name{validate_SE}
\alias{validate_SE}
\title{Validate SummarizedExperiment object}
\usage{
validate_SE(se)
}
\arguments{
\item{se}{SummarizedExperiment object 
produced by the gDR pipeline}
}
\value{

}
\description{
Validate SummarizedExperiment object
}
