% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{write_config}
\alias{write_config}
\title{Function that saves the configuration to the yaml file}
\usage{
write_config(config, path)
}
\arguments{
\item{config}{configuration "object"}

\item{path}{path to the output yaml file}
}
\value{
TRUE on success, FALSE on error
}
\description{
Function that saves the configuration to the yaml file
}
