% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatentate_SEs.R
\name{aggregate_assay}
\alias{aggregate_assay}
\title{Aggregate a \code{BumpyMatrix} assay by a given aggreation function.}
\usage{
aggregate_assay(asy, aggregation_fxn, by)
}
\arguments{
\item{asy}{A \code{BumpyMatrix} object.}
}
\value{
A \code{BumpyMatrix} object aggregated by \code{aggregation_fxn}.
}
\description{
Aggregation can only be performed on nested variables.
}
