% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{assay_to_dt}
\alias{assay_to_dt}
\title{Transform a SummarizedExperiment assay to a long data.table}
\usage{
assay_to_dt(se, assay_name, merge_metrics = FALSE, include_metadata = TRUE)
}
\arguments{
\item{se}{\linkS4class{SummarizedExperiment} object with dose-response data.}

\item{assay_name}{String of name of the assay or index of the assay in the \code{se}.}

\item{merge_metrics}{Logical indicating whether the metrics should be merged.
Defaults to \code{FALSE}.}

\item{include_metadata}{Boolean indicating whether to include the metadata on the SummarizedExperiment.
Defaults to \code{TRUE}.}
}
\value{
data.table with dose-response data
}
\description{
Transform a SummarizedExperiment assay to a long data.table with a single entry for each row and column combination.
}
