% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_choices}
\alias{assert_choices}
\title{assert choices}
\usage{
assert_choices(x, choices, ...)
}
\arguments{
\item{x}{charvec expected subset}

\item{choices}{charvec reference set}
}
\description{
assert choices
}
