% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_biological_replicates_dt}
\alias{average_biological_replicates_dt}
\title{Average biological replicates.}
\usage{
average_biological_replicates_dt(
  dt,
  var,
  pidfs = get_prettified_identifiers(),
  fixed = TRUE,
  geometric_average_fields = get_header("metric_average_fields")$geometric_mean
)
}
\arguments{
\item{dt}{data.table with Metric data}

\item{fixed}{Flag should be add fix for -Inf in geometric mean.}

\item{geometric_average_fields}{Character vector of column names in \code{dt}
to take the geometric average of.}

\item{average_fields}{Character vector of column names in \code{dt} to take
the average of.}
}
\value{
data.table without replicates
}
\description{
Average biological replicates on the data table side.
}
