% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_to_dt.R
\name{convert_assay_data_to_dt.BumpyDataFrameMatrix}
\alias{convert_assay_data_to_dt.BumpyDataFrameMatrix}
\title{Convert assay data into data.table}
\usage{
\method{convert_assay_data_to_dt}{BumpyDataFrameMatrix}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Convert assay data into data.table
}
