% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_to_dt.R
\name{convert_se_ref_assay_to_dt}
\alias{convert_se_ref_assay_to_dt}
\title{Convert the reference values from a SummarizedExperiment assay to a long data.table}
\usage{
convert_se_ref_assay_to_dt(se)
}
\arguments{
\item{se}{A \linkS4class{SummarizedExperiment} object holding raw and/or processed dose-response data in its assays.}
}
\value{
data.table representation of the reference data.
}
\description{
Convert the Ref[RelativeViability/GRvalue] within a \linkS4class{SummarizedExperiment} object to a long data.table. Clean up the column names and add columns to match the format of the data.table from the 'Averaged' assay.
}
\details{
NOTE: to extract information about 'Control' data, simply call the 
function with the name of the assay holding data on controls.
}
