% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_normalization_type_name}
\alias{extend_normalization_type_name}
\title{extend abbreviated normalization type}
\usage{
extend_normalization_type_name(x)

extend_normalization_type_name(x)
}
\arguments{
\item{x}{string with normalization type}
}
\value{
string

string
}
\description{
extend abbreviated normalization type

extend abbreviated normalization type
}
\examples{
extend_normalization_type_name("GR")

}
