% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{fit_curves}
\alias{fit_curves}
\title{Fit curves}
\usage{
fit_curves(
  df_,
  e_0 = 1,
  GR_0 = 1,
  n_point_cutoff = 4,
  range_conc = c(0.005, 5),
  force_fit = FALSE,
  pcutoff = 0.05
)
}
\arguments{
\item{df_}{data.frame containing data to fit. See details.}

\item{e_0}{numeric value representing the \code{x_0} value for the RV curve.
Defaults to \code{1}.}

\item{GR_0}{numeric value representing the \code{x_0} value for the GR curve.
Defaults to \code{1}.}

\item{n_point_cutoff}{integer of how many points should be considered the minimum required to try to fit a curve.
Defaults to \code{4}.}

\item{range_conc}{numeric vector of length 2 indicating the lower and upper concentration ranges.
Defaults to \code{c(5e-3, 5)}. See details.}

\item{force_fit}{boolean indicating whether or not to force a constant fit.
Defaults to \code{FALSE}.}

\item{pcutoff}{numeric of pvalue significance threshold above which to use a constant fit.
Defaults to \code{0.05}.}
}
\value{
data.frame of RV and GR fit parameters.
}
\description{
Fit GR and RV curves from a data.frame.
}
\details{
The \code{df_} expects the following columns:
 \itemize{
  \item{Concentration} unlogged concentration values
  \item{RelativeViability} normalized relative viability values
  \item{std_RelativeViability} standard relative viability values
  \item{GRvalue} normalized GR values
  \item{std_GRvalue} standard GR values
 }

The \code{range_conc} is used to calculate the \code{x_AOC_range} statistic.
The purpose of this statistic is to enable comparison across different experiments with slightly 
different concentration ranges.
}
