% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRutils-package.R
\docType{package}
\name{gDRutils-package}
\alias{gDRutils}
\alias{gDRutils-package}
\title{gDRutils: A package with helper functions for processing drug response data}
\value{
package help page
}
\description{
This package contains utility functions used throughout the gDR platform to fit data, manipulate data, and convert and validate data structures. This package also has the necessary default constants for gDR platform. Many of the functions are utilized by the gDRcore package.
}
\note{
To learn more about functions start with \code{help(package = "gDRutils")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDRutils}
  \item \url{https://gdrplatform.github.io/gDRutils/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDRutils/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
  \item Aleksander Chlebowski
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Pawel Piatkowski
  \item Dariusz Scigocki
  \item Janina Smola
  \item Sergiu Mocanu
  \item Allison Vuong
}

}
\keyword{internal}
