% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRutils-package.R
\docType{package}
\name{gDRutils-package}
\alias{gDRutils}
\alias{gDRutils-package}
\title{gDRutils: A package with helper functions for processing drug response data}
\value{
package help page
}
\description{
This package contains utility functions used throughout the gDR platform. Contaion many helper functions for maniplating data, including e.g. converting and validating. This package also has the necessary default constants for gDR platform.
}
\note{
To learn more about functions start with \code{help(package = "gDRutils")}
}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com}

Authors:
\itemize{
  \item Bartosz Czech
  \item Aleksander Chlebowski
  \item Marc Hafner
  \item Pawel Piatkowski
  \item Dariusz Scigocki
  \item Sergiu Mocanu
  \item Allison Vuong
}

}
\keyword{internal}
