% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_metadata.R
\name{get_MAE_identifiers}
\alias{get_MAE_identifiers}
\title{get_MAE_identifiers}
\usage{
get_MAE_identifiers(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment}
}
\value{
named list with identifiers for each SE
}
\description{
get the identifiers of all SE's in the MAE
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
get_MAE_identifiers(mae)

}
