% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_non_empty_assays}
\alias{get_non_empty_assays}
\title{get_non_empty_assays}
\usage{
get_non_empty_assays(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
charvec with non-empty experiments
}
\description{
get non empty assays
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
get_non_empty_assays(mae)

}
\author{
Arkadiusz Gladki \href{mailto:arkadiusz.gladki@contractors.roche.com}{arkadiusz.gladki@contractors.roche.com}
}
