% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{get_optional_coldata_fields}
\alias{get_optional_coldata_fields}
\title{get optional colData fields}
\usage{
get_optional_coldata_fields(se)
}
\arguments{
\item{se}{a SummarizedExperiment object with drug-response data generate by gDR pipeline}
}
\value{
a charvec containing the names of the optional identifiers in the SE colData
}
\description{
get optional colData fields
}
\keyword{standardize_MAE}
