% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{get_supported_experiments}
\alias{get_supported_experiments}
\title{get_supported_experiments}
\usage{
get_supported_experiments(type = NULL)
}
\arguments{
\item{type}{String indicating the type of experiment}
}
\value{
charvec with supported experiment name(s)
}
\description{
get supported experiments
}
\examples{
get_supported_experiments()

}
\author{
Arkadiusz Gladki \href{mailto:arkadiusz.gladki@contractors.roche.com}{arkadiusz.gladki@contractors.roche.com}
}
\keyword{experiment}
