% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{identifiers}
\alias{identifiers}
\alias{get_identifier}
\alias{set_identifier}
\alias{reset_identifiers}
\title{Get, set, or reset identifiers for one or all identifier field(s)}
\usage{
get_identifier(k = NULL)

set_identifier(k, v)

reset_identifiers()
}
\arguments{
\item{k}{string of field to get or set identifiers for}

\item{v}{character vector of identifiers corresponding to identifier field \code{k}}
}
\value{
For \code{set_identifier} or \code{reset_identifiers} a \code{NULL} invisibly.
For \code{get_identifier} a character vector of identifiers for field \code{k}.
}
\description{
Get, set, or reset the expected identifier(s) for one or all identifier field(s)
}
\details{
If \code{get_identifier} is called with no arguments, the entire available identifier list is returned.

\code{set_identifier} can be called on an identifier that is already set.
A common use case to do so is when trying to load multiple files of different identifiers
in the same R session.

\code{reset_identifiers} should be used if processing multiple files within the same R session
that use different file identifiers.
}
\examples{
\dontrun{
get_identifier("duration")
set_identifier("duration", "SET DURATION TIME")
get_identifier("duration")
reset_identifiers() 
get_identifier("duration")
}
}
