% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_SE.R
\name{identify_unique_se_metadata_fields}
\alias{identify_unique_se_metadata_fields}
\title{Indetify unique se metadata fields from the list of SEs}
\usage{
identify_unique_se_metadata_fields(SElist)
}
\arguments{
\item{SElist}{named list of Summarized Experiments}
}
\value{
vector of unique names of metadata
}
\description{
Indetify unique se metadata fields from the list of SEs
}
