% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_SE.R
\name{identify_unique_se_metadata_fields}
\alias{identify_unique_se_metadata_fields}
\title{Identify unique metadata fields from a list of \code{SummarizedExperiment}s}
\usage{
identify_unique_se_metadata_fields(SElist)
}
\arguments{
\item{SElist}{named list of \code{SummarizedExperiment}s}
}
\value{
character vector of unique names of metadata
}
\description{
Identify unique metadata fields from a list of \code{SummarizedExperiment}s
}
\examples{
SElist <- list(
  get_synthetic_data("finalSE_small.RDS"), 
  get_synthetic_data("finalSE_small_no_noise.RDS")
)
identify_unique_se_metadata_fields(SElist)

}
