% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_any_exp_empty}
\alias{is_any_exp_empty}
\title{is_any_exp_empty}
\usage{
is_any_exp_empty(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
logical
}
\description{
check if any experiment is empty
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
is_any_exp_empty(mae)

}
\author{
Arkadiusz Gladki \href{mailto:arkadiusz.gladki@contractors.roche.com}{arkadiusz.gladki@contractors.roche.com}
}
\keyword{package_utils}
