% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_exp_empty}
\alias{is_exp_empty}
\title{is_exp_empty}
\usage{
is_exp_empty(exp)
}
\arguments{
\item{exp}{\linkS4class{SummarizedExperiment} object.}
}
\value{
logical
}
\description{
check if experiment (SE) is empty
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
se <- mae[[1]]
is_exp_empty(se)

}
\author{
Arkadiusz Gladki \href{mailto:arkadiusz.gladki@contractors.roche.com}{arkadiusz.gladki@contractors.roche.com}
}
\keyword{package_utils}
