% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_mae_empty}
\alias{is_mae_empty}
\title{is_mae_empty}
\usage{
is_mae_empty(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
logical
}
\description{
check if all mae experiments are empty
}
\author{
Arkadiusz Gladki <arkadiusz.gladki@contractors.roche.com>
}
