% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_SE.R
\name{merge_assay}
\alias{merge_assay}
\title{Merge assay data}
\usage{
merge_assay(
  SElist,
  assay_name,
  discard_keys = NULL,
  additional_col_name = NULL,
  returnDT = TRUE
)
}
\arguments{
\item{SElist}{named list of Summarized Experiments}

\item{assay_name}{name of the assay that should be extracted and merged}

\item{discard_keys}{a character vector of string that will be discarded
during creating BumpyMatrix object}

\item{additional_col_name}{a string with the name of the column that will be
added to assay data for the distinction of possible duplicated metrics
that can arise from multiple projects}

\item{returnDT}{a logical indication whether a list of BM and DT should be returned}
}
\value{
BumpyMatrix or list with data.table + BumpyMatrix
}
\description{
Merge assay data
}
