% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mrowData}
\alias{mrowData}
\title{mrowData}
\usage{
mrowData(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
tibble with all-experiments rowData
}
\description{
get rowData of all experiments
}
\author{
Arkadiusz Gladki <arkadiusz.gladki@contractors.roche.com>
}
