% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{predict_smooth_from_combo}
\alias{predict_smooth_from_combo}
\title{Predict a smoothed response for a drug combination}
\usage{
predict_smooth_from_combo(conc_1, conc_2, metrics_merged)
}
\arguments{
\item{conc_1}{A single numeric value for the desired concentration of the first drug.}

\item{conc_2}{A single numeric value for the desired concentration of the second drug.}

\item{metrics_merged}{A data.table containing all pre-calculated curve fit parameters.
Expects columns: 'dilution_drug', 'cotrt_value', 'ratio', 'ec50', 'h', 'x_inf', 'x_0'.}
}
\value{
A single numeric value for the predicted 'smooth' response.
}
\description{
Predicts a 'smooth' value for a single pair of drug concentrations by
snapping to the nearest available models from a metrics table and averaging
their predictions. This is the combination equivalent of
'predict_efficacy_from_conc'.
}
\examples{
metrics <- data.table::data.table(
 dilution_drug = c("drug_1", "drug_1", "drug_2", "drug_2"),
 cotrt_value = c(0, 10, 0, 1),
 ratio = c(NA, NA, NA, NA),
 ec50 = c(1, 1.5, 5, 6),
 h = c(2, 2, 2, 2),
 x_inf = c(0.1, 0.2, 0.1, 0.3),
 x_0 = c(1, 1, 1, 1)
)
predict_smooth_from_combo(conc_1 = 1.2, conc_2 = 9.8, metrics_merged = metrics)

}
\keyword{fit_curves}
