% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_assay_to_dt.R
\name{prettify_flat_metrics}
\alias{prettify_flat_metrics}
\title{Prettify metric names in flat 'Metrics' assay}
\usage{
prettify_flat_metrics(
  x,
  human_readable = FALSE,
  normalization_type = c("GR", "RV")
)
}
\arguments{
\item{x}{a list of names.}

\item{human_readable}{boolean indicating whether or not to return column names in human readable format.
Defaults to \code{FALSE}.}

\item{normalization_type}{a character with a specified normalization type.
Defaults to \code{c("GR", "RV")}.}
}
\value{
character vector of prettified names.
}
\description{
Map existing column names of a flattened 'Metrics' assay to prettified names.
}
\details{
Rename names that are metrics in a table.

A common use case for this function is to convert column names from a flattened version of the \code{"Metrics"} assay.
Mode \code{"human_readable = TRUE"} is often used in the context of front-end applications,
whereas \code{"human_readable" = FALSE} is often used in the context of the command line.
}
