% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_se_assay_to_dt.R
\name{prettify_flat_metrics}
\alias{prettify_flat_metrics}
\title{Pretify metric names in flat 'Metrics' assay}
\usage{
prettify_flat_metrics(name_list, human_readable = FALSE)
}
\arguments{
\item{name_list}{a list of names.}

\item{human_readable}{either \code{"variable"} (default) or \code{"gDRviz"}.}
}
\value{
table of data with updated column names.
}
\description{
Return pretified names for flat 'Metrics' assay
}
\details{
Rename names that are  metrics in a table.

A common use case for this function is to convert column names from a flattened version of the \code{"Metrics"} assay.
}
