% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{refine_rowdata}
\alias{refine_rowdata}
\title{refine rowData}
\usage{
refine_rowdata(rd, se, default_v = "Undefined")
}
\arguments{
\item{rd}{DataFrame with colData}

\item{se}{a SummarizedExperiment object with drug-response data generate by gDR pipeline}

\item{default_v}{string with default value for optional columns in rowData}
}
\value{
refined rowData
}
\description{
current improvements done on the rowData as a standardization step:
- set default value for optional rowData fields
}
