% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{rename_bumpy}
\alias{rename_bumpy}
\title{Rename BumpyMatrix}
\usage{
rename_bumpy(bumpy, mapping_vector)
}
\arguments{
\item{bumpy}{a BumpyMatrix object}

\item{mapping_vector}{a named vector for mapping old and new values.
The names of the character vector indicate the source names, and the
corresponding values the destination names.}
}
\value{
a renamed BumpyMatrix object
}
\description{
Rename BumpyMatrix
}
\examples{
mae <- get_synthetic_data("finalMAE_small.qs")
se <- mae[[1]]
assay <- SummarizedExperiment::assay(se)
rename_bumpy(assay, c("Concentration" = "conc"))

}
