% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{send_email}
\alias{send_email}
\title{send email by wrapping \code{mailR::send.mail}}
\usage{
send_email(
  body,
  subject,
  to = get_env_var("EMAIL_RECIPIENT"),
  from = get_env_var("EMAIL_SENDER"),
  html = TRUE,
  inline = FALSE,
  host_name = get_env_var("EMAIL_HOSTNAME"),
  attached_files = NULL
)
}
\arguments{
\item{body}{a character containing the body content of the email}

\item{subject}{A string with the subject of the email}

\item{to}{a string with the recipient's email address}

\item{from}{string with the sender's email address}

\item{html}{a logical flag indicating whether the body of the email should be parsed as HTML}

\item{inline}{a logical flag indicating whether images in the HTML file should be embedded inline}

\item{host_name}{a string with the hostname}

\item{attached_files}{a character with file paths to be attached to the email}
}
\description{
send email by wrapping \code{mailR::send.mail}
}
\keyword{utils}
