% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{set_identifier}
\alias{set_identifier}
\title{Deprecated function, use set_env_identifier.}
\usage{
set_identifier(k, v)
}
\arguments{
\item{k}{String corresponding to identifier name.}

\item{v}{Character vector corresponding to the value for given identifier \code{k}.}
}
\description{
Deprecated function, use set_env_identifier.
}
