% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{set_unique_cl_names_dt}
\alias{set_unique_cl_names_dt}
\title{Set Unique Parental Identifiers in table}
\usage{
set_unique_cl_names_dt(col_data, sep = " ")
}
\arguments{
\item{col_data}{data.table or DFrame with col data}

\item{sep}{string with separator added before suffix}
}
\value{
fixed input table with unique \code{CellLineName} in \code{colData}.
}
\description{
This function sets the \code{CellLineName} field in
\code{colData} to be unique by appending the \code{clid} in parentheses for duplicates.
}
\examples{
col_data <- S4Vectors::DataFrame(CellLineName = c("ID1", "ID1"), clid = c("C1", "C2"))
col_data <- set_unique_cl_names_dt(col_data)
}
\keyword{standardize_MAE}
