% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_MAE.R
\name{set_unique_drug_names_dt}
\alias{set_unique_drug_names_dt}
\title{Set Unique Drug Names in table}
\usage{
set_unique_drug_names_dt(row_data, sep = " ")
}
\arguments{
\item{row_data}{data.table or DFrame with row data}

\item{sep}{string with separator added before suffix}
}
\value{
fixed input table with unique \code{DrugName} fields in \code{rowData}.
}
\description{
This function sets the \code{DrugName}, \code{DrugName_2}, and \code{DrugName_3} fields in
\code{rowData} to be unique by appending the corresponding \code{Gnumber}, \code{Gnumber_2},
and \code{Gnumber_3} in parentheses for duplicates.
}
\examples{
row_data <- S4Vectors::DataFrame(
  DrugName = c("DrugA", "DrugA", "DrugB"),
  Gnumber = c("G1", "G2", "G5"),
  DrugName_2 = c("DrugC", "DrugC", "DrugD"),
  Gnumber_2 = c("G3", "G4", "G5")
)
row_data <- set_unique_drug_names_dt(row_data)
}
\keyword{standardize_MAE}
