% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_convert.R
\name{strip_first_and_last_char}
\alias{strip_first_and_last_char}
\title{String first and last characters of a string.}
\usage{
strip_first_and_last_char(jstring)
}
\arguments{
\item{jstring}{String of any number of characters greater than 1.}
}
\value{
String with first and last characters stripped.
}
\description{
String first and last characters of a string.
}
\details{
This is most often used to remove the JSON brackets \code{'{'} and \code{'}'}.
}
\keyword{json_convert}
