% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{update_idfs_synonyms}
\alias{update_idfs_synonyms}
\title{Update gDR synonyms for the identifier}
\usage{
update_idfs_synonyms(data, dict = get_idfs_synonyms())
}
\arguments{
\item{data}{list of charvec with identifiers data}

\item{dict}{list with dictionary}
}
\value{
list
}
\description{
Update gDR synonyms for the identifier
}
\examples{
mdict <- list(duration = "time")
iv <- c("Time", Duration, "time")
update_idfs_synonyms(iv, dict = mdict)

}
