% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_validators.R
\name{validate_MAE}
\alias{validate_MAE}
\title{Validate MultiAssayExperiment object}
\usage{
validate_MAE(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object 
produced by the gDR pipeline}
}
\value{
\code{NULL} invisibly if the MultiAssayExperiment is valid.
Throws an error if the MultiAssayExperiment is not valid.
}
\description{
Function validates correctness of SE included in MAE by checking multiple cases:
- detection of duplicated rowData/colData,
- incompatibility of rownames/colnames,
- occurrence of necessary assays,
- detection of mismatch of CLIDs inside colData and colnames (different order),
- correctness of metadata names.
}
\author{
Bartosz Czech <bartosz.czech@contractors.roche.com>
}
