% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_validators.R
\name{validate_SE}
\alias{validate_SE}
\title{Validate SummarizedExperiment object}
\usage{
validate_SE(se)
}
\arguments{
\item{se}{SummarizedExperiment object 
produced by the gDR pipeline}
}
\value{
\code{NULL} invisibly if the assay name is valid.
Throws an error if the SummarizedExperiment is not valid.
}
\description{
Function validates correctness of SE by checking multiple cases:
- detection of duplicated rowData/colData,
- incompatibility of rownames/colnames,
- occurrence of necessary assays,
- detection of mismatch of CLIDs inside colData and colnames (different order),
- correctness of metadata names.
}
